osa4 done

+++ osa4/controllers/blogs.js
@@ -1,20 +1,41 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const logger = require('../utils/logger')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+    const body = request.body
+    const user = request.user
+
+    const blog = new Blog({
+        ...body,
+        user: user._id
+    })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
     const id = request.params.id
+    const user = request.user
+    const blog = await Blog.findById(id)
+
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if (blog.user.toString() !== user._id.toString()) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
     await Blog.findByIdAndRemove(id)
     response.status(204).end()
 })
